Added capabilities for managing the details returned from the queries.  Specificlly offered options for filtering the list of details returned based on a new parameter called "key_filter" and an option to only return results that match a given "result_filter".  These new parameters are optional so any existing use of the code should continue to work.
Updates to both the library and command line utitlies was made to support these new capabilities.  For the CLI utilities, only a single key/value pair for searching for certain results is supported, while the library supports any number of key/value pairs.  This can be fixed in the utility in the future but will require work on the argparsing to improve the capability.